Improved checkstyle integration tests - Don't overload the well-behaved plugin test - Actually execute checkstyle when configured by configuration
